+++ Code/Engine/System/Basics.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <Foundation/Basics.h>
+#include <System/PlatformFeatures.h>
 
 // Configure the DLL Import/Export Define
 #if EZ_ENABLED(EZ_COMPILE_ENGINE_AS_DLL)
@@ -15,4 +16,6 @@
   #define EZ_SYSTEM_DLL
   #define EZ_SYSTEM_TEMPLATE
 #endif
+
+