+++ Projects/SampleApp_Game/Window.h
@@ -16,12 +16,6 @@ private:
   void CreateContextOGL();
   void DestroyContextOGL();
 
-#if EZ_ENABLED(EZ_PLATFORM_WINDOWS)
-  void OnWindowMessage(HWND hWnd, UINT Msg, WPARAM WParam, LPARAM LParam);
-#else
-  #error "No window input handling!"
-#endif
-
   HDC m_hDC;
   HGLRC m_hRC;
 };